requirements.txt
mss>=9.0.0
Pillow>=10.0.0
keyboard>=0.13.5
pystray>=0.19.0
pywin32>=306
winotify>=1.1.0
